Move Patch Logic
Prompt: "Move the apply_patch and patch_files functions out of main.py and into patch.py. Make sure to add imports to main.py to be able to call the moved functions."
